Diagnostic HTML and CSS
Comfort: 4
Clarity: 4

I used Google and my class notes to help with this assessment.
I also used the HTTP study repo from me.

I am not very comfortable with the positioning. I find it very hard to create
a wireframe. It would be helpful for me to go through the task from yesterday
once together so that I can see hoe to approach such kind of problem. Now it's
a lot of trying around.